apps/consumer/pyproject.toml
[tool.poetry]
name = "consumer"
version = "0.1.0"
description = ""
authors = ["Mark Shteyn <markshteyn1@gmail.com>"]
readme = "README.md"
package-mode=false

[tool.poetry.dependencies]
python = ">=3.11,<3.14"
fastapi = "^0.115.5"
uvicorn = "^0.32.1"
pika = "^1.3.2"
google-api-python-client = "^2.154.0"
pytesseract = "^0.3.13"
pandas = "^2.2.3"
pillow = "^11.0.0"
pypdf2 = "^3.0.1"
docx = "^0.2.4"
mammoth = "^1.8.0"
striprtf = "^0.0.28"
bs4 = "^0.0.2"
python-docx = "^1.1.2"
pymupdf = "^1.25.0"
odf = "^0.0.1"
odfpy = "^1.4.1"
python-dotenv = "^1.0.1"
weaviate-client = "^4.10.1"
openai = "^1.57.4"
requests = "^2.32.3"
aio-pika = "^9.5.4"
olefile = "^0.47"
cairosvg = "^2.7.1"
prisma = "^0.15.0"
pydantic = "^2.10.4"
python-multipart = "^0.0.20"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
